+++ CHANGELOG.md
@@ -36,6 +36,8 @@
 ### Fixed
 
 - [#5512](https://github.com/ChainSafe/forest/pull/5512) Fixed `Filecoin.EthTraceFilter` RPC method.
+- [#5517](https://github.com/ChainSafe/forest/issues/5517) Fix the `forest-cli sync wait` issue
+- [#5540](https://github.com/ChainSafe/forest/pull/5540) Avoid duplicate migrations.
 
 ## Forest v0.25.2 "Fool of a Took"